fix(DDH-14): build fail
