+++ src/components/index.js
@@ -1,4 +1,3 @@
 export * from "./AppLayout";
 export * from "./MapLabel";
-export * from "./Map";
 export * from "./filter";
